feat: App 코드 분리 및 useMemo 사용과 Context 용도별 분리
