+++ src/AppProviders.tsx
@@ -1,0 +1,50 @@
+// AppProviders.tsx
+import React, { useState, useMemo } from "react";
+import { ThemeContext, UserContext, NotificationContext } from "./contexts";
+import { User, Notification } from "./types";
+
+export const AppProviders: React.FC<{ children: React.ReactNode }> = ({
+  children,
+}) => {
+  const [theme, setTheme] = useState("light");
+  const [user, setUser] = useState<User | null>(null);
+  const [notifications, setNotifications] = useState<Notification[]>([]);
+
+  const toggleTheme = () =>
+    setTheme((prev) => (prev === "light" ? "dark" : "light"));
+
+  const login = (email: string) => {
+    setUser({ id: 1, name: "홍길동", email });
+    addNotification("성공적으로 로그인되었습니다", "success");
+  };
+
+  const logout = () => {
+    setUser(null);
+    addNotification("로그아웃되었습니다", "info");
+  };
+
+  const addNotification = (message: string, type: Notification["type"]) => {
+    const newNotification = { id: Date.now(), message, type };
+    setNotifications((prev) => [...prev, newNotification]);
+  };
+
+  const removeNotification = (id: number) =>
+    setNotifications((prev) => prev.filter((n) => n.id !== id));
+
+  const themeContextValue = useMemo(() => ({ theme, toggleTheme }), [theme]);
+  const userContextValue = useMemo(() => ({ user, login, logout }), [user]);
+  const notificationContextValue = useMemo(
+    () => ({ notifications, addNotification, removeNotification }),
+    [notifications]
+  );
+
+  return (
+    <ThemeContext.Provider value={themeContextValue}>
+      <UserContext.Provider value={userContextValue}>
+        <NotificationContext.Provider value={notificationContextValue}>
+          {children}
+        </NotificationContext.Provider>
+      </UserContext.Provider>
+    </ThemeContext.Provider>
+  );
+};
