+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./Header";
+export * from "./ItemList";
+export * from "./ComplexForm";
